Create a new refactored ListAdapter for media items. Moreover, create an extension function that helps to get all visible items
