+++ app/src/main/java/com/yurii/youtubemusic/utilities/Extensions.kt
@@ -1,0 +1,19 @@
+package com.yurii.youtubemusic.utilities
+
+import androidx.recyclerview.widget.RecyclerView
+
+fun <T : RecyclerView.ViewHolder> RecyclerView.getVisibleItems(): List<T> {
+    val arrayList = ArrayList<T>()
+    for (index: Int in 0 until childCount) {
+        val child = getChildAt(index)
+        val position = getChildAdapterPosition(child)
+
+        if (position == RecyclerView.NO_POSITION)
+            continue
+
+        val viewHolder = getChildViewHolder(child) as T
+        arrayList.add(viewHolder)
+    }
+
+    return arrayList
+}
